Refactor purchase form
